+++ business/src/main/java/com/kevinguanchedarias/owgejava/repository/CriticalAttackRepository.java
@@ -1,0 +1,8 @@
+package com.kevinguanchedarias.owgejava.repository;
+
+import com.kevinguanchedarias.owgejava.entity.CriticalAttack;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CriticalAttackRepository extends JpaRepository<CriticalAttack, Integer> {
+
+}