+++ game-frontend/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 v0.10.0 (latest)
 =================
+
+* [class=Admin] __Feature:__ [As an admin I want to add critical attack to certain unit types or units #266](https://github.com/KevinGuancheDarias/owge/issues/266)  
+* [class=Admin] __Fix:__ Unit attack rule not closing afer saving
 * __Fix:__ [When an error occur in the sync endpoint, the sync breaks completely, bug due to #386 ....  #390](https://github.com/KevinGuancheDarias/owge/issues/390)  
 * __Feature:__ [As a player I want the game to have a multiaccount detection system #380](https://github.com/KevinGuancheDarias/owge/issues/380)  
 * __Fix:__ [When a conquest succeed, is not deleting the building mission #389](https://github.com/KevinGuancheDarias/owge/issues/389)